+++ src/main/java/com/gregtechceu/gtceu/client/util/DrawUtil.java
@@ -1,6 +1,6 @@
 package com.gregtechceu.gtceu.client.util;
 
-import com.gregtechceu.gtceu.core.mixins.GuiGraphicsAccessor;
+import com.gregtechceu.gtceu.core.mixins.client.GuiGraphicsAccessor;
 
 import com.lowdragmc.lowdraglib.utils.ColorUtils;
