+++ src/main/java/com/gregtechceu/gtceu/core/mixins/client/GuiHeartTypeMixin.java
@@ -1,4 +1,4 @@
-package com.gregtechceu.gtceu.core.mixins;
+package com.gregtechceu.gtceu.core.mixins.client;
 
 import com.gregtechceu.gtceu.common.data.GTMobEffects;
