+++ src/main/java/com/gregtechceu/gtceu/core/mixins/client/HumanoidArmorLayerMixin.java
@@ -1,4 +1,4 @@
-package com.gregtechceu.gtceu.core.mixins;
+package com.gregtechceu.gtceu.core.mixins.client;
 
 import com.gregtechceu.gtceu.common.item.armor.GTArmorItem;
 
@@ -28,7 +28,8 @@ public abstract class HumanoidArmorLayerMixin<T extends LivingEntity, M extends 
     @ModifyArg(method = "renderArmorPiece",
                at = @At(value = "INVOKE",
                         target = "Lnet/minecraft/client/renderer/entity/layers/HumanoidArmorLayer;renderModel" +
-                                "(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;ILnet/minecraft/world/item/ArmorItem;Lnet/minecraft/client/model/Model;ZFFFLnet/minecraft/resources/ResourceLocation;)V"),
+                                "(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;ILnet/minecraft/world/item/ArmorItem;Lnet/minecraft/client/model/Model;ZFFFLnet/minecraft/resources/ResourceLocation;)V",
+                        remap = false),
                index = 6)
     private float gtceu$modifyArmorTintR(float oldR, @Local ArmorItem armorItem) {
         if (armorItem instanceof GTArmorItem gtArmorItem) {
@@ -47,7 +48,8 @@ public abstract class HumanoidArmorLayerMixin<T extends LivingEntity, M extends 
     @ModifyArg(method = "renderArmorPiece",
                at = @At(value = "INVOKE",
                         target = "Lnet/minecraft/client/renderer/entity/layers/HumanoidArmorLayer;renderModel" +
-                                "(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;ILnet/minecraft/world/item/ArmorItem;Lnet/minecraft/client/model/Model;ZFFFLnet/minecraft/resources/ResourceLocation;)V"),
+                                "(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;ILnet/minecraft/world/item/ArmorItem;Lnet/minecraft/client/model/Model;ZFFFLnet/minecraft/resources/ResourceLocation;)V",
+                        remap = false),
                index = 7)
     private float gtceu$modifyArmorTintsG(float oldG, @Local ArmorItem armorItem) {
         if (armorItem instanceof GTArmorItem gtArmorItem) {
@@ -66,7 +68,8 @@ public abstract class HumanoidArmorLayerMixin<T extends LivingEntity, M extends 
     @ModifyArg(method = "renderArmorPiece",
                at = @At(value = "INVOKE",
                         target = "Lnet/minecraft/client/renderer/entity/layers/HumanoidArmorLayer;renderModel" +
-                                "(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;ILnet/minecraft/world/item/ArmorItem;Lnet/minecraft/client/model/Model;ZFFFLnet/minecraft/resources/ResourceLocation;)V"),
+                                "(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;ILnet/minecraft/world/item/ArmorItem;Lnet/minecraft/client/model/Model;ZFFFLnet/minecraft/resources/ResourceLocation;)V",
+                        remap = false),
                index = 8)
     private float gtceu$modifyArmorTintsB(float oldB, @Local ArmorItem armorItem) {
         if (armorItem instanceof GTArmorItem gtArmorItem) {
