+++ src/main/java/com/gregtechceu/gtceu/core/mixins/client/ModelManagerMixin.java
@@ -1,4 +1,4 @@
-package com.gregtechceu.gtceu.core.mixins;
+package com.gregtechceu.gtceu.core.mixins.client;
 
 import com.gregtechceu.gtceu.GTCEu;
 import com.gregtechceu.gtceu.client.renderer.block.MaterialBlockRenderer;
