+++ src/main/resources/gtceu.mixins.json
@@ -6,23 +6,23 @@
     "compatibilityLevel": "JAVA_17",
     "plugin": "com.gregtechceu.gtceu.core.mixins.GTMixinPlugin",
     "client": [
-        "AbstractClientPlayerAccessor",
-        "BiomeColorsMixin",
-        "BlockModelMixin",
-        "ClientLevelAccessor",
-        "GuiGraphicsAccessor",
-        "GuiGraphicsMixin",
-        "GuiHeartTypeMixin",
-        "HumanoidArmorLayerMixin",
-        "LevelRendererMixin",
-        "ModelManagerMixin",
-        "MultiPlayerGameModeMixin",
-        "PlayerInfoAccessor",
-        "VariantDeserializerMixin",
+        "client.AbstractClientPlayerAccessor",
+        "client.BiomeColorsMixin",
+        "client.BlockModelMixin",
+        "client.ClientLevelAccessor",
+        "client.GuiGraphicsAccessor",
+        "client.GuiGraphicsMixin",
+        "client.GuiHeartTypeMixin",
+        "client.HumanoidArmorLayerMixin",
+        "client.LevelRendererMixin",
+        "client.ModelManagerMixin",
+        "client.MultiPlayerGameModeMixin",
+        "client.PlayerInfoAccessor",
+        "client.VariantDeserializerMixin",
         "ftbchunks.FTBChunksClientMixin",
         "ftbchunks.LargeMapScreenMixin",
         "ftbchunks.RegionMapPanelMixin",
-        "ldlib.CustomBakedModelAccessor",
+        "ldlib.ModularWrapperWidgetMixin",
         "rei.FluidEntryRendererMixin",
         "xaerominimap.HighlighterRegistryMixin",
         "xaerominimap.MinimapFBORendererMixin",
@@ -66,7 +66,6 @@
         "forge.ConfiguredModelListAccessor",
         "forge.ModelBuilderAccessor",
         "jei.FluidHelperMixin",
-        "ldlib.ModularWrapperWidgetMixin",
         "ldlib.SyncUtilsMixin",
         "rei.InputSlotCrafterMixin",
         "rei.RecipeFinderMixin",
